// src/components/textField/Input.d.ts
import React from 'react';
import { ForwardRefInjectedProps } from '../../commons/new';
import { InputProps } from './types';
declare const Input: {
    ({ flex, left, top, right, bottom, style, hint, color, forwardedRef, formatter, useGestureHandlerInput, readonly, recorderTag, pointerEvents, showMandatoryIndication, ...props }: InputProps & ForwardRefInjectedProps): React.JSX.Element;
    displayName: string;
};
export default Input;