// src/components/textField/ValidationMessage.d.ts
import React from 'react';
import { ValidationMessageProps } from './types';
declare const ValidationMessage: {
    ({ validationMessage, validationIcon, enableErrors, validationMessageStyle, retainValidationSpace, validate, testID }: ValidationMessageProps): React.JSX.Element | null;
    displayName: string;
};
export default ValidationMessage;