// src/components/textField/Label.d.ts
import React from 'react';
import { LabelProps } from './types';
declare const Label: {
    ({ label, labelColor, labelStyle, labelProps, validationMessagePosition, floatingPlaceholder, showMandatoryIndication, enableErrors, testID }: LabelProps): React.JSX.Element | null;
    displayName: string;
};
export default Label;